[RUI-241]: Move getScatterChartProOptions function to index file
